requirements.txt
pytest==8.2.2
playwright==1.46.0
pytest-xdist==3.6.1
python-dotenv==1.0.1
allure-pytest==2.13.5
requests==2.32.3
sentence-transformers==2.7.0
torch>=2.1.0
